fix(#273): unify AgentId as u64 across QuarantineRegistry, reconciler, and all boundaries
Promotes `phantom-agents::agent::AgentId` from `u32` to `u64`, matching
`phantom-agents::role::AgentId` and `QuarantineRegistry`. Removes the
`u32::try_from(agent_id).unwrap_or(u32::MAX)` saturating cast in the
reconciler (and the accompanying `agent_id as u32` cast in check_stalled).

Changes across the stack:
- phantom-agents/src/agent.rs: AgentId = u64 (was u32)
- phantom-brain/src/orchestrator.rs: PlanStep::agent_id Option<u64>
- phantom-brain/src/reconciler.rs: remove both narrowing casts; rewrite
  the #272 saturation test as a #273 round-trip correctness test
- phantom-protocol/src/events.rs: AgentId = u64 (IPC wire type)
- phantom-app/src/adapters/agent.rs: widen app_id to u64 at wire boundary
- phantom-app/src/agent_pane.rs: remove stale `as u64` casts, update comment
- phantom-session/src/{agent_state,goal_state}.rs: widen persistence types
- phantom-agents/src/failure_store.rs: fix test range literals (u32→u64)
- phantom-brain/src/brain.rs: fix test literal types (u32→u64)

cargo check --workspace passes; phantom-agents and phantom-brain test
suites pass (605 + 343 tests, 0 failures).

+++ crates/phantom-agents/src/agent.rs
@@ -17,7 +17,11 @@ use crate::tools::{ToolCall, ToolResult};
 // ---------------------------------------------------------------------------
 
 /// Unique agent identifier (monotonically increasing within a session).
-pub type AgentId = u32;
+///
+/// Widened to `u64` (matches [`crate::role::AgentId`] and
+/// [`crate::quarantine::QuarantineRegistry`]) so no narrowing cast is required
+/// at the dispatch / quarantine boundary. Fixes issue #273.
+pub type AgentId = u64;
 
 // ---------------------------------------------------------------------------
 // AgentStatus
